+++ NEWS.md
@@ -1,5 +1,7 @@
 # chevron 0.2.5.9005
 
+* Soft deprecated `strat` argument of `kmg01_main` - use `strata` instead.
+
 # chevron 0.2.5
 
 * `MNG01` plot can now be displayed without error bars and can display a continuous temporal scale on the `x` axis.